pyproject.toml
[project]
name = "henri"
version = "0.1.0"
description = "Humanitarian Early-warning Network Resilience Intelligence"
requires-python = ">=3.11"
dependencies = [
    "playwright>=1.40",
    "httpx>=0.25",
    "pandas>=2.1",
    "pyarrow>=14.0",
    "click>=8.1",
    "pycountry>=23.12",
    "python-dotenv>=1.0",
    "jinja2>=3.1",
    "thefuzz[speedup]>=0.20",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "apscheduler>=3.10,<4",
]

[project.scripts]
henri = "cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend/snow_extract", "backend/snow_parse", "backend/snow_analyse", "backend/grafana_client", "backend/netbox_client", "backend/osint", "backend/henri"]

[tool.hatch.build.targets.wheel.sources]
"backend" = ""

[tool.pytest.ini_options]
pythonpath = ["backend"]

[dependency-groups]
dev = [
    "pip-audit>=2.10.0",
    "pytest>=9.0.2",
]
